Re-saved images for web, saved people images for web, removed unused images, began styling CSS.
I saved out new, full versions of the works images, as well as saving
for web all of the Our Team images. I deleted the responsive images
that were being used for Exercise Four, I deleted all of the red class
divs, I reordered the work section as well as the team section, and I
began  stylizing text and placing in imagery.